requirements.txt
zeroc-ice
pymongo